% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.marketplacecommerceanalytics_package.R
\docType{package}
\name{paws.marketplacecommerceanalytics-package}
\alias{paws.marketplacecommerceanalytics}
\alias{paws.marketplacecommerceanalytics-package}
\title{paws.marketplacecommerceanalytics: AWS Marketplace Commerce Analytics}
\description{
Provides AWS Marketplace business intelligence data
    on-demand.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
